# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Book Management API
  description: |
    RESTful API for the Book Management system built with Node.js, Express, and MySQL.
    Provides authentication, user profile management, book catalog operations, PDF report generation, and document uploads.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000
    description: Local Development Server
...
paths:
...
  /books:
    get:
      summary: Get all books
      description: Retrieves a list of all books, with optional filtering by author name.
      tags:
        - Books
      parameters:
        - $ref: '#/components/parameters/AuthorQueryParam'
      responses:
        '200':
          description: A list of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        '500':
          $ref: '#/components/responses/ErrorResponse'

    post:
      summary: Add a new book
      description: Creates a new book entry in the database. Requires authentication.
      tags:
        - Books
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreateRequest'
      responses:
        '201':
          description: Book successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookCreatedResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Please provide title, author, and price
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ErrorResponse'
...
  /documents:
    get:
      summary: Get all uploaded documents metadata
      description: Returns a list of metadata for all uploaded documents (id, title, filename, filesize).
      tags:
        - Documents
      responses:
        '200':
          description: List of document metadata
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentMetadata'
        '500':
          $ref: '#/components/responses/ErrorResponse'

    post:
      summary: Upload a document
      description: Uploads a single PDF document (maximum file size 20MB) along with a title.
      tags:
        - Documents
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
                - file
              properties:
                title:
                  type: string
                  description: Title or label for the document
                  example: Programming Fundamentals Module 1
                file:
                  type: string
                  format: binary
                  description: PDF file to upload (Max 20MB)
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentUploadResponse'
        '400':
          description: Invalid file format (only PDF allowed) or size limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Only PDF files are allowed
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Provide your JSON Web Token (JWT) retrieved from `/auth/login`.
        Header format: `Authorization: Bearer <token>`
  parameters:
...
    AuthorQueryParam:
      name: author
      in: query
      required: false
      description: Filter books by author name
      schema:
        type: string
      example: Robert C. Martin
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: Operation completed successfully
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Error description
...
    Book:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Clean Code
        author:
          type: string
          example: Robert C. Martin
        price:
          type: number
          format: float
          example: 29.99
        genre:
          type: string
          nullable: true
          example: Programming
        published_year:
          type: integer
          nullable: true
          example: 2008
    BookCreateRequest:
      type: object
      required:
        - title
        - author
        - price
      properties:
        title:
          type: string
          example: The Pragmatic Programmer
        author:
          type: string
          example: Andrew Hunt, David Thomas
        price:
          type: number
          format: float
          example: 45.50
...
    BookCreatedResponse:
      type: object
      properties:
        message:
          type: string
          example: Book added
        bookId:
          type: integer
          example: 1
    DocumentMetadata:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Lecture 12 Notes
        filename:
          type: string
          example: 1789292111056-SE1102_ProgrammingFundamentals_L12M1.pdf
        filesize:
          type: integer
          example: 2048576
    MulterFile:
      type: object
      properties:
        fieldname:
          type: string
          example: file
        originalname:
          type: string
          example: SE1102_ProgrammingFundamentals_L12M1.pdf
        encoding:
          type: string
          example: 7bit
        mimetype:
          type: string
          example: application/pdf
        destination:
          type: string
          example: uploads/
        filename:
          type: string
          example: 1789292111056-SE1102_ProgrammingFundamentals_L12M1.pdf
        path:
          type: string
          example: uploads/1789292111056-SE1102_ProgrammingFundamentals_L12M1.pdf
        size:
          type: integer
          example: 2048576
    DocumentUploadResponse:
      type: object
      properties:
        message:
          type: string
          example: File uploaded successfully
        file:
          $ref: '#/components/schemas/MulterFile'
  responses:
    UnauthorizedError:
      description: Authentication token is missing, invalid, or expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
          examples:
            missingToken:
              summary: Missing authentication token
              value:
                message: Authentication token required
            invalidToken:
              summary: Invalid or expired token
              value:
                message: Invalid or expired token
    ErrorResponse:
      description: Internal server error with error details
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'